Move createMockNodeRED in a general file
