+++ test/unit/email-receiver.unit.test.js
@@ -1,10 +1,13 @@
 const should = require('should');
 const {
-  createMockNodeRED,
   setupModuleMocks,
   testConfigs,
   testUtils
 } = require('../helpers/email-receiver.mocks.js');
+const {
+    createMockNodeRED
+} = require('../helpers/general.mocks.js');
+
 
 describe('Email Receiver Node - Unit Tests with Helpers', function() {
   this.timeout(10000);
